+++ src/commands/group.ts
@@ -6,7 +6,7 @@ import { TaskCommand } from '@/src/modules/task-command.tsx';
 import { LoadProject } from '@/src/middlewares/load-project.ts';
 import { LazyCurrentWorkspace, LoadWorkspace } from '@/src/middlewares/load-workspace.ts';
 import { type Workspace, type WorkspaceDepsMode } from '@/src/project/workspace.ts';
-import { TaskExprService, type TaskTree } from '@/src/tasks/task-expr.service.ts';
+import { TaskExpressionService, type TaskTree } from '@/src/tasks/task-expression.service.ts';
 
 // Types
 export interface IGroupCommandArgs {
@@ -32,8 +32,8 @@ export class GroupCommand extends TaskCommand<IGroupCommandArgs> {
 
   // Constructor
   constructor(
-    @inject(TaskExprService)
-    private readonly taskExpr: TaskExprService,
+    @inject(TaskExpressionService)
+    private readonly taskExpr: TaskExpressionService,
   ) {
     super();
   }
